requirements.txt
flask
opencv-python
Pillow
numpy
python-dotenv
pandas
matplotlib
seaborn
requests
torch==1.10.1
torchvision==0.11.2
pyyaml
tqdm